Add files via upload
Cut 1 of Hacker Mode for Alexa(tm).  This project is an Alexa skill and requires some work on your part to run it on your Alexa. I will publish versions of this publicly with public-friendly skills. Versions that do local scanning will probably be kept private for contributors when we get to that point.
The two directories AlexaSkill and Lambda indicate which of Amazon's web sites handle the data contained in each. (Yep you have to log into 2 web sites to complete the Hacker Mode app. Speech data in one, program in the other... you'll get used to it. READ THE README!!!