Fixed Lic
Added Amazon's required Use Limitation, and my use limitation based on
protecting the project as a collaborative effort and the name of the
project's invocation phrase. This should make the project safe to work
on as a group without another user cloning the project and naming it
something similar and registering it with Amazon and other troll
behavior.